+++ jgiven-core/src/main/java/com/tngtech/jgiven/report/model/DataTable.java
@@ -3,6 +3,7 @@ package com.tngtech.jgiven.report.model;
 import java.util.List;
 
 import com.tngtech.jgiven.annotation.Table.HeaderType;
+import com.tngtech.jgiven.impl.util.AssertionUtil;
 
 /**
  * Represents a data table argument.
@@ -15,7 +16,7 @@ public class DataTable {
     private HeaderType headerType;
 
     /**
-     * The data of the table.
+     * The data of the table as a list of rows
      */
     private List<List<String>> data;
 
@@ -40,6 +41,10 @@ public class DataTable {
         this.data = data;
     }
 
+    public int getRowCount() {
+        return data.size();
+    }
+
     @Override
     public boolean equals( Object o ) {
         if( this == o )
@@ -62,5 +67,33 @@ public class DataTable {
         int result = headerType != null ? headerType.hashCode() : 0;
         result = 31 * result + ( data != null ? data.hashCode() : 0 );
         return result;
+    }
+
+    public boolean hasHorizontalHeader() {
+        return headerType == HeaderType.HORIZONTAL || headerType == HeaderType.BOTH;
+    }
+
+    public boolean hasVerticalHeader() {
+        return headerType == HeaderType.VERTICAL || headerType == HeaderType.BOTH;
+    }
+
+    public void addColumn( int i, List<String> column ) {
+        AssertionUtil.assertTrue( data.size() == column.size(),
+            "Column has different number of rows as expected. Is " + column.size() + ", but expected " + data.size() );
+
+        for( int row = 0; row < column.size(); row++ ) {
+            data.get( row ).add( i, column.get( row ) );
+        }
+    }
+
+    public int getColumnCount() {
+        return data.get( 0 ).size();
+    }
+
+    public void addRow( int i, List<String> row ) {
+        AssertionUtil.assertTrue( getColumnCount() == row.size(),
+            "Row has different number of columns as expected. Is " + row.size() + ", but expected " + getColumnCount() );
+
+        data.add( i, row );
     }
 }